005_chatbot_firestore_flask/requirements.txt
wheel
grpcio
google-cloud-firestore
flask
flask_cors
line-bot-sdk